study/requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-filter==22.1
django-import-export==3.0.1
django-taggit==3.0.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.38.0
future==0.18.2
html5lib==1.1
idna==3.4
lxml==4.9.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
Pillow==9.3.0
pycparser==2.21
pydyf==0.5.0
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyPDF3==1.0.6
pyphen==0.13.0
python-bidi==0.4.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
svglib==1.4.1
tablib==3.2.1
tinycss2==1.2.1
tqdm==4.64.1
tzdata==2022.6
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
webencodings==0.5.1
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.1
